The refactor is complete and websocket is working again.
This change is so all modules have access to the websocket and session info 🚀
