+++ oidc-controller/api/main.py
@@ -8,12 +8,12 @@ from api.core.config import settings
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 from fastapi.middleware.cors import CORSMiddleware
-import socketio # For using websockets
 
 from .routers import acapy_handler, oidc, presentation_request, well_known_oid_config
 from .verificationConfigs.router import router as ver_configs_router
 from .clientConfigurations.router import router as client_config_router
 from .db.session import init_db, get_db
+from .routers.socketio import sio_app
 
 from api.core.oidc.provider import init_provider
 
@@ -52,36 +52,8 @@ app.include_router(
     oidc.router, prefix="/vc/connect", tags=["oidc-deprecated"], include_in_schema=False
 )
 
-##################################
-# Configure the websocket
-# TODO: Move this to a separate file
-sio = socketio.AsyncServer(async_mode='asgi', cors_allowed_origins='*')
-
-@sio.event
-async def connect(sid, socket):
-    print('connected', sid)
-    # logger.info(f">>> connect : sid={sid}")
-    # logger.info(f">>> connect : socket={socket}")
-    await sio.emit('message', {'data': "I'm a real boy!"})
-    # TODO: Add the sid 
-
-@sio.event
-async def initialize(sid, data):
-    logger.info(f">>> initialize : sid={sid}")
-    logger.info(f">>> initialize : pid={data.get('pid')}")
-    # TODO: Add the pid to the acapy_handler.connections dict
-
-@sio.event
-async def disconnect(sid):
-    logger.info(f">>> disconnect : sid={sid}")
-    # TODO: Remove the sid,pid & socket from the acapy_handler.connections dict
-
-##################################
-# TODO: Keep this here and import sio
-sio_app = socketio.ASGIApp(socketio_server=sio)
-
+# Connect the websocket server to run within the FastAPI app
 app.mount('/ws', sio_app)
-##################################
 
 origins = ["*"]
